Fix type errors after explicitly typing operations
