+++ waspc/src/Wasp/Generator/SdkGenerator/Server/OperationsGenerator.hs
@@ -16,7 +16,7 @@ import Wasp.AppSpec.Operation (getName)
 import qualified Wasp.AppSpec.Operation as AS.Operation
 import qualified Wasp.AppSpec.Query as AS.Query
 import Wasp.AppSpec.Valid (isAuthEnabled)
-import Wasp.Generator.Common (makeJsonWithEntityData)
+import Wasp.Generator.Common (dropExtensionFromImportPath, makeJsonWithEntityData)
 import Wasp.Generator.FileDraft (FileDraft)
 import qualified Wasp.Generator.JsImport as GJI
 import Wasp.Generator.Monad (Generator)
@@ -158,6 +158,6 @@ extImportToJsImport extImport@(EI.ExtImport extImportName extImportPath) =
       _importAlias = Just $ EI.importIdentifier extImport ++ "_ext"
     }
   where
-    importPath = C.makeSdkImportPath $ extCodeDirP </> SP.castRel extImportPath
+    importPath = C.makeSdkImportPath $ dropExtensionFromImportPath $ extCodeDirP </> SP.castRel extImportPath
     extCodeDirP = fromJust $ SP.relDirToPosix C.extCodeDirInSdkRootDir
     importName = GJI.extImportNameToJsImportName extImportName
